docker/backend/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.3.0
pymongo==4.1.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
